+++ CHANGELOG.md
@@ -452,3 +452,4 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 * Custom save fields for cards and relics (twanvl)
 
 #### dev ####
+* Make Spriter animation public (kiooeht)